front: move SuggestedOP type in modules/timetableItem/types.ts
