+++ front/src/modules/pathfinding/hooks/usePathfinding.ts
@@ -20,7 +20,7 @@ import {
   getPathfindingQuery,
   matchPathStepAndOp,
 } from 'modules/pathfinding/utils';
-import type { SuggestedOP } from 'modules/timetableItem/components/ManageTimetableItem/types';
+import type { SuggestedOP } from 'modules/timetableItem/types';
 import { setFailure, setWarning } from 'reducers/main';
 import { replaceItinerary, updatePathSteps } from 'reducers/osrdconf/operationalStudiesConf';
 import {
