+++ front/src/modules/pathfinding/utils.ts
@@ -11,7 +11,7 @@ import type {
   RollingStockWithLiveries,
 } from 'common/api/osrdEditoastApi';
 import { getSupportedElectrification, isThermal } from 'modules/rollingStock/helpers/electric';
-import type { SuggestedOP } from 'modules/timetableItem/components/ManageTimetableItem/types';
+import type { SuggestedOP } from 'modules/timetableItem/types';
 import type { PathStep } from 'reducers/osrdconf/types';
 import { addElementAtIndex } from 'utils/array';
 import { getPointOnTrackCoordinates } from 'utils/geometry';
