+++ front/src/modules/timesStops/types.ts
@@ -1,6 +1,6 @@
 import type { TrainSchedule } from 'common/api/osrdEditoastApi';
 import type { TimeString } from 'common/types';
-import type { SuggestedOP } from 'modules/timetableItem/components/ManageTimetableItem/types';
+import type { SuggestedOP } from 'modules/timetableItem/types';
 import type { ArrayElement } from 'utils/types';
 
 export type TimeExtraDays = {
