+++ front/src/modules/timetableItem/components/ManageTimetableItem/ManageTimetableItemMap/ItineraryMarkers.tsx
@@ -15,7 +15,7 @@ import type { PathItemLocation, TrackSection } from 'common/api/osrdEditoastApi'
 import { matchPathStepAndOp } from 'modules/pathfinding/utils';
 import type { PathStep } from 'reducers/osrdconf/types';
 
-import type { SuggestedOP } from '../types';
+import type { SuggestedOP } from '../../../types';
 
 export type MarkerInformation = Pick<PathStep, 'name' | 'coordinates' | 'metadata'> &
   PathItemLocation & {
